feat: refactor MusicProvidersLoginContent component structure and update import path
